Move composer reducers together
